% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actual_RV_gen.R
\docType{data}
\name{actual_RV_gen}
\alias{actual_RV_gen}
\title{Actual Solar Generation Data}
\format{
A data frame with 6720 rows and 6 variables:
\describe{
\item{REGIONID}{character, region identifier}
\item{INTERVAL_DATETIME}{character, datetime of each power generation interval}
\item{DATE}{character, date of the power generation record}
\item{TIME}{character, time of the power generation record}
\item{POWER}{numeric, recorded power generation in megawatts}
}
}
\usage{
actual_RV_gen
}
\description{
A dataset containing solar generation power data across regions.
}
\keyword{datasets}
